qr alignment
Fix: Adjust QR code and text alignment on the display

This commit slightly adjusts the x-coordinate of the "Waiting for IP address..." text and the QR code to improve their horizontal alignment on the display. The text is moved from x: 100 to x: 90, and the QR code is moved from x: -5 to x: 0. These changes result in a more visually centered appearance.